requirements.txt
matplotlib
torch
transformers
opencv-python-headless
scipy
Pillow
panel
bokeh
numpy<2
scikit-image
